Reflow

+++ src/backend/src/resolvers/mutation.js
@@ -1,20 +1,116 @@
+const bcrypt = require('bcryptjs')
+const jwt = require('jsonwebtoken')
+
 const { db } = require('../db')
 const { sanitize } = require('../sanitize')
 
 const Mutation = {
+  async register (parent, args, ctx, info) {
+    // Set email to lowercase to avoid duplication
+    args.email = args.email.toLowerCase()
+
+    // Hash password
+    const password = await bcrypt.hash(args.password, 16)
+
+    const user = await db.mutation.createUser(
+      {
+        data: { ...args, password, permissions: { set: ['USER'] } }
+      },
+      info
+    )
+
+    // Create the JWT
+    const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET)
+
+    // Store the cookie with the user for one (1) year
+    ctx.res.cookie('token', token, {
+      httpOnly: true,
+      maxAge: 1000 * 3600 * 24 * 365
+    })
+
+    return user
+  },
+
+  async login (parent, args, ctx, info) {
+    // Set email to lowercase to avoid duplication
+    const email = args.email.toLowerCase()
+
+    // Find user by email
+    const user = await db.query.user({ where: { email } })
+
+    if (!user) {
+      throw Error(`No such user found for this email: ${email}`)
+    }
+
+    // Confirm the password
+    const isValid = await bcrypt.compare(args.password, user.password)
+
+    if (!isValid) {
+      throw Error('Invalid password')
+    }
+
+    // Create token for user
+    const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET)
+
+    // Store the cookie with the user for one (1) year
+    ctx.res.cookie('token', token, {
+      httpOnly: true,
+      maxAge: 1000 * 3600 * 24 * 365
+    })
+
+    return user
+  },
+
+  logout (parent, args, ctx, info) {
+    ctx.res.clearCookie('token')
+
+    return 'Goodbye!'
+  },
+
   async createAdventure (parent, args, ctx, info) {
+    const { userId } = ctx.req
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
     const { title, description } = args
 
     const cleanDescription = sanitize(description)
 
     return db.mutation.createAdventure(
-      { data: { title, description: cleanDescription } },
+      {
+        data: {
+          title,
+          description: cleanDescription,
+          owner: { connect: { id: userId } }
+        }
+      },
       info
     )
   },
 
   async updateAdventure (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { id, ...updates } = args
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    // Check if allowed to update adventure
+    const adventure = await db.query.adventure(
+      { where: { id } },
+      `{ owner { id } }`
+    )
+
+    if (!adventure) {
+      return null
+    } else if (adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
+    }
 
     if (updates.description) {
       updates.description = sanitize(updates.description)
@@ -30,18 +126,47 @@ const Mutation = {
   },
 
   async deleteAdventure (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { id } = args
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    // Check if allowed to delete adventure
+    const adventure = await db.query.adventure(
+      { where: { id } },
+      `{ owner { id } }`
+    )
+
+    if (!adventure) {
+      return null
+    } else if (adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
+    }
 
     return db.mutation.deleteAdventure({ where: { id } }, info)
   },
 
   async createSession (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { adventureId, title, description } = args
 
-    const adventure = await db.query.adventure({ where: { id: adventureId } })
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    const adventure = await db.query.adventure(
+      { where: { id: adventureId } },
+      `{ owner { id } }`
+    )
 
     if (!adventure) {
       throw Error(`No such Adventure for ${adventureId}`)
+    } else if (adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
     }
 
     const cleanDescription = sanitize(description)
@@ -59,7 +184,24 @@ const Mutation = {
   },
 
   async updateSession (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { id, ...updates } = args
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    const session = await db.query.session(
+      { where: { id } },
+      `{ adventure { owner { id } } }`
+    )
+
+    if (!session) {
+      return null
+    } else if (session.adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
+    }
 
     if (updates.description) {
       updates.description = sanitize(updates.description)
@@ -69,18 +211,46 @@ const Mutation = {
   },
 
   async deleteSession (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { id } = args
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    const session = await db.query.session(
+      { where: { id } },
+      `{ adventure { owner { id } } }`
+    )
+
+    if (!session) {
+      return null
+    } else if (session.adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
+    }
 
     return db.mutation.deleteSession({ where: { id } }, info)
   },
 
   async createQuest (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { adventureId, title, description, completed = false } = args
 
-    const adventure = await db.query.adventure({ where: { id: adventureId } })
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    const adventure = await db.query.adventure(
+      { where: { id: adventureId } },
+      `{ owner { id }  }`
+    )
 
     if (!adventure) {
       throw Error(`No such Adventure for ${adventureId}`)
+    } else if (adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
     }
 
     const cleanDescription = sanitize(description)
@@ -99,7 +269,24 @@ const Mutation = {
   },
 
   async updateQuest (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { id, ...updates } = args
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    const quest = await db.query.quest(
+      { where: { id } },
+      `{ adventure { owner { id } } }`
+    )
+
+    if (!quest) {
+      return null
+    } else if (quest.adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
+    }
 
     if (updates.description) {
       updates.description = sanitize(updates.description)
@@ -109,7 +296,24 @@ const Mutation = {
   },
 
   async deleteQuest (parent, args, ctx, info) {
+    const { userId } = ctx.req
     const { id } = args
+
+    // Check for user ID
+    if (!userId) {
+      throw Error('You must be logged in to do that!')
+    }
+
+    const quest = await db.query.quest(
+      { where: { id } },
+      `{ adventure { owner { id } } }`
+    )
+
+    if (!quest) {
+      return null
+    } else if (quest.adventure.owner.id !== userId) {
+      throw Error(`You don't have permission to do that!`)
+    }
 
     return db.mutation.deleteQuest({ where: { id } }, info)
   }
